Question Paper Uploaded
To download the Question paper either you can do by Downloading the Zip File or Clone it by clicking on the "Fork" button on the top right corner then copying the url from the "Green Button" and then in Git terminal write command : - ( git clone "url you copied" )

You have to attempt any one question paper and name the Repository of your own account as given in the Rule Book. After uploading project in the repository share the Link of that repository on WhatsApp Group of Xordium.

Happy Coding!